Split app shell orchestration hooks
